+++ tests/testthat/test-s3-methods.R
@@ -1,4 +1,5 @@
 context("S3")
+set.seed(5)
 n <- 10
 dat <- data.frame(
   x = rep(0:1, times = 5),
@@ -187,7 +188,7 @@ test_that("tidy, glance, summary, and print work", {
     mpg ~ cyl + am,
     data = mtcars,
     se_type = "classical",
-    linear_hypothesis = c("cyl = am", "cyl = 3")
+    linear_hypothesis = "cyl = am"
   )
   tlho <- tidy(lho)
   slho <- summary(lho)
@@ -226,7 +227,7 @@ test_that("tidy, glance, summary, and print work", {
   # tidy adds rows for each LH
   expect_equal(
     tlho$term,
-    c("(Intercept)", "cyl", "am", "cyl = am", "cyl = 3")
+    c("(Intercept)", "cyl", "am", "cyl = am")
   )
 
   # glance only glances lm_robust object
@@ -246,19 +247,19 @@ test_that("tidy, glance, summary, and print work", {
   )
 
   expect_identical(
-    printsummary_lho[19:21],
+    printsummary_lho[19:20],
     capture.output(summary(lho$lh))
   )
 
   expect_output(
     print(lho$lh),
-    "Estimate.*cyl = am.*cyl = 3"
+    "Estimate.*cyl = am"
   )
 
   # print also gets right number of rows
   expect_equal(
     length(capture.output(print(lho$lh))),
-    3
+    2
   )
 
   ## horvitz_thompson
@@ -641,6 +642,59 @@ test_that("predict works", {
   expect_equivalent(
     predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
     predict(lm_int_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
+  )
+
+  # With and without factorial treatment, intercept, predictions should be the same
+  lml_out <- lm_lin(y ~ x, covariates = ~ z + I(z^2) +cl, data = dat, se_type = "classical")
+  lml_0_out <- lm_lin(y ~ x + 0, covariates = ~ z + I(z^2) + cl, data = dat, se_type = "classical")
+  lml_f_out <- lm_lin(y ~ factor(x), covariates = ~ z + I(z^2) + cl, data = dat, se_type = "classical")
+  lm_f0_out <- lm_lin(y ~ factor(x) + 0, covariates = ~ z + I(z^2) + cl, data = dat, se_type = "classical")
+
+  expect_equivalent(
+    predict(lml_0_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
+  )
+  expect_equivalent(
+    predict(lml_f_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+  )
+
+  expect_equivalent(
+    predict(lml_f_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
+  )
+
+  expect_equivalent(
+    predict(lml_f_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lm_f0_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
+  )
+
+  # For multi-level treatments, with and without factorial treatment, intercept, predictions should be the same
+  dat$x_multi <- rep(0:2, times = 7)[1:20]
+  new_dat$x_multi <- rep(0:2, times = 7)
+
+  lml_out <- lm_lin(y ~ x_multi, covariates = ~ z + I(z^2) +cl, data = dat, se_type = "classical")
+  lml_0_out <- lm_lin(y ~ x_multi + 0, covariates = ~ z + I(z^2) + cl, data = dat, se_type = "classical")
+  lml_f_out <- lm_lin(y ~ factor(x_multi), covariates = ~ z + I(z^2) + cl, data = dat, se_type = "classical")
+  lm_f0_out <- lm_lin(y ~ factor(x_multi) + 0, covariates = ~ z + I(z^2) + cl, data = dat, se_type = "classical")
+
+  expect_equivalent(
+    predict(lml_0_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
+  )
+  expect_equivalent(
+    predict(lml_f_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+  )
+
+  expect_equivalent(
+    predict(lml_f_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lml_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
+  )
+
+  expect_equivalent(
+    predict(lml_f_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)],
+    predict(lm_f0_out, new_dat, se.fit = TRUE, interval = "confidence")[c(1, 2)]
   )
 
   # working with rank deficient X